Use linked resource instead of filesystem
Use linked resources instead of filesystem hack as this one can return
invalid locations through standard API.

This removes the filesystem bundle, the logic is now moved in
ProjectManagers
and a workspace listener takes care of using linked resources for
metadata.

This also remove the
InvisibleProjectMetadataTest.testMetadataFileLocation() test because
creating a new project in Eclipse workspace now enforces creation of a
.settings/org.eclipse.core.resources.prefs file. Where the metadata for
the invisible project is stored doesn't matter as the internal project
is internal and not visible to user anyway.